+++ lib/index.js
@@ -2,7 +2,7 @@
 const _ = require('lodash');
 const BbPromise = require('bluebird');
 const packageParameters = require('./package/parameters/packageParameters');
-const packageMaintenanceWindow = require('./package/parameters/packageMaintenanceWindow');
+const packageMaintenanceWindow = require('./package/maintenanceWindows/packageMaintenanceWindow');
 const packageDocuments= require('./package/documents/packageDocuments');
 const validations= require('./utils/validations');
