+++ lib/index.test.js
@@ -2,14 +2,14 @@
 
 const ServerlessSystemsManager = require('./index');
 const packageParameters = require('./package/parameters/packageParameters');
-const packageMaintenanceWindow = require('./package/parameters/packageMaintenanceWindow');
+const packageMaintenanceWindow = require('./package/maintenanceWindows/packageMaintenanceWindow');
 const packageDocuments = require('./package/documents/packageDocuments');
 
 
 jest.mock('./package/parameters/packageParameters', () => ({
   packageParameters: jest.fn().mockResolvedValue('mockedValue'),
 }));
-jest.mock('./package/parameters/packageMaintenanceWindow', () => ({
+jest.mock('./package/maintenanceWindows/packageMaintenanceWindow', () => ({
   packageMaintenanceWindow: jest.fn().mockResolvedValue('mockedValue'),
 }));
